+++ src/components/post/PostCard/index.jsx
@@ -1,22 +1,19 @@
 import React, { useState, useEffect } from 'react';
-import { deleteDoc, doc, getDoc, setDoc, updateDoc } from 'firebase/firestore';
 import { v4 as uuidv4 } from 'uuid';
 import { useLocation, useNavigate } from 'react-router';
 import * as S from './style';
-import { db } from '../../../firebase';
 import IconHeartOn from '../../../assets/Icon-Heart-on.png';
 import IconHeartOff from '../../../assets/Icon-Heart-off.png';
 import IconStarOn from '../../../assets/Icon-star-on.png';
 import IconStarOff from '../../../assets/Icon-star-off.png';
 import useToggle from '../../../hooks/useToggle';
-import Portal from '../../modal/Portal';
-import ConfirmModal from '../../modal/ConfirmModal';
+import useLikeUpdate from '../../../hooks/useLikeUpdate';
 
 function PostCard({ id, date, like, location, menu, photo, review, score, shop, tags, postList }) {
   const scoreIndexs = [0, 1, 2, 3, 4];
   const [liked, setLiked] = useToggle(like);
   const [formattedDate, setFormattedDate] = useState();
-  const [isConfirmModalOpen, setIsConfirmModalOpen] = useState(false);
+
   const navigate = useNavigate();
   const { pathname } = useLocation();
 
@@ -33,37 +30,33 @@ function PostCard({ id, date, like, location, menu, photo, review, score, shop, 
     formatDate(dateFormatted);
   }, []);
 
-  const updatePost = async (postId, newLiked) => {
-    const postDoc = doc(db, 'post', postId);
-    const newField = { like: newLiked };
+  // const updatePost = async (postId, newLiked) => {
+  //   const postDoc = doc(db, 'post', postId);
+  //   const newField = { like: newLiked };
 
-    await updateDoc(postDoc, newField);
+  //   await updateDoc(postDoc, newField);
 
-    if (newLiked) {
-      const postData = (await getDoc(postDoc)).data();
+  //   if (newLiked) {
+  //     const postData = (await getDoc(postDoc)).data();
 
-      await setDoc(doc(db, 'liked', id), {
-        ...postData,
-        like: newLiked,
-      });
-    } else {
-      await deleteDoc(doc(db, 'liked', id));
-    }
-  };
+  //     await setDoc(doc(db, 'liked', id), {
+  //       ...postData,
+  //       like: newLiked,
+  //     });
+  //   } else {
+  //     await deleteDoc(doc(db, 'liked', id));
+  //   }
+  // };
+
+  useLikeUpdate(id, liked);
 
   const handleLikeButton = (e) => {
     if (pathname === '/likeposts') {
       e.stopPropagation();
-      setIsConfirmModalOpen(true);
     } else {
       setLiked(!liked);
-      updatePost(id, !liked);
       e.stopPropagation();
     }
-  };
-
-  const confirmModalClose = () => {
-    setIsConfirmModalOpen(false);
   };
 
   const handleClickCard = () => {
@@ -115,20 +108,6 @@ function PostCard({ id, date, like, location, menu, photo, review, score, shop, 
           </S.PostReview>
         </S.PostContent>
       </S.PostCardBox>
-      <Portal>
-        <ConfirmModal
-          visible={isConfirmModalOpen}
-          msg='좋아요 목록에서 삭제할까요?'
-          leftBtnMsg='취소'
-          rightBtnMsg='삭제'
-          onClickClose={confirmModalClose}
-          rightOnclick={() => {
-            updatePost(id, !liked);
-            setIsConfirmModalOpen((prev) => !prev);
-          }}
-          leftOnclick={confirmModalClose}
-        />
-      </Portal>
     </>
   );
 }
